# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "PluginPlay can help you setup Google Cloud Platform (GCP) in minutes. It can spin up new compute instances, configure them and monitor them. Before running any command, we want the user to confirm given a draft of commands that will run",
      "version": "1.0.0"
    },
    "servers": [
        {
          "url": "http://localhost:8000"
        }
      ],
    "paths": {
        "/list-compute-instances": {
            "get": {
                "operationId": "listComputeInstances",
                "summary": "List Google Cloud Platform Compute Engine instances",
                "description": "Retrieve a list of Compute Engine instances in a specified zone.",
              "responses": {
                "200": {
                  "description": "A JSON object containing an array of GCP instance names.",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "instances": {
                            "type": "array",
                            "instance": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Bad request. The request is malformed or invalid."
                },
                "500": {
                  "description": "Internal server error. An error occurred while processing the request."
                }
              }
            }
          },
          "/create-compute-instance": {
            "post": {
              "operationId": "createComputeInstance",
              "summary": "Create a new GCP Compute Instance",
              "description": "Creates a new GCP Compute Instance with the specified instance name",
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "instance_name": {
                          "type": "string",
                          "description": "The name of the GCP Compute Instance to be created"
                        }
                      },
                      "required": [
                        "instance_name"
                      ]
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "Instance created successfully and returns a instance URL, that users can click on to view their instance",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "instance_url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Bad request (e.g., missing or invalid input)",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/configure-compute-instance": {
            "post": {
              "operationId": "configureComputeInstance",
              "summary": "Configure the GCP compute instance by running a list of commands. Before running any command, we want the user to confirm given a draft of commands that will run",
              "description": "This endpoint accepts an array of commands and executes them on the Google Cloud Platform (GCP) compute instance. Before running any command, we want the user to confirm given a draft of commands that will run",
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "instance_name": {
                          "type": "string",
                          "description": "The name of the GCP compute instance on which to execute the commands"
                        },
                        "commands": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "An array of commands to be executed on the GCP compute instance"
                        }
                      },
                      "required": [
                        "instance_name",
                        "commands"
                      ]
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "Commands ran successfully"
                },
                "400": {
                  "description": "Bad request",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "description": "A message describing the error."
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "description": "A message describing the error."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/visualize-setup": {
            "get": {
              "operationId": "visualizeSetup",
              "summary": "Visualize the GCP setup. Don't just give me the link, rather render the video by writing it in Markdown without backticks and without using a code block",
              "description": "Get a visual representation of the GCP setup. Don't just give me the link, rather render the video by writing it in Markdown without backticks and without using a code block. This video has to render",
              "responses": {
                "200": {
                  "description": "A URL to an video file representing the GCP setup",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "visualization_url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Bad request",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "Image not found",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
    }
  }